// src/app/proto/s2c.d.ts
export namespace S2C {

    interface IMessage {
        SC_TEST_ECHO?: (S2C.ISC_TEST_ECHO|null);
        Error?: (S2C.IError|null);
        LOGIN_SC_LOGIN?: (S2C.ILOGIN_SC_LOGIN|null);
        LOGIN_SC_CHOOSE_SERVER?: (S2C.ILOGIN_SC_CHOOSE_SERVER|null);
        LOGIN_SC_GET_SERVER_LIST?: (S2C.ILOGIN_SC_GET_SERVER_LIST|null);
        LOGIN_SC_GET_INFO?: (S2C.ILOGIN_SC_GET_INFO|null);
        LOGIN_SC_QUERY?: (S2C.ILOGIN_SC_QUERY|null);
        SC_ROLE_CREATE?: (S2C.ISC_ROLE_CREATE|null);
        SC_ROLE_ONLINE?: (S2C.ISC_ROLE_ONLINE|null);
        SC_ROLE_HEART_BEAT?: (S2C.ISC_ROLE_HEART_BEAT|null);
        SC_UPDATE_ROLE_PRO?: (S2C.ISC_UPDATE_ROLE_PRO|null);
        SC_GET_REWARD?: (S2C.ISC_GET_REWARD|null);
        SC_GM_CMD?: (S2C.ISC_GM_CMD|null);
        SC_UPDATE_RECHARGE?: (S2C.ISC_UPDATE_RECHARGE|null);
        SC_ROLE_OFFLINE?: (S2C.ISC_ROLE_OFFLINE|null);
    }
// ...
    interface ISC_ROLE_HEART_BEAT {
        msg?: (string|null);
        clientVersion?: (string|null);
    }
// ...
    interface ILOGIN_SC_LOGIN {
        serverId?: (number|null);
        gmAuth?: (number|null);
        isNew?: (number|null);
    }
// ...
    interface ILOGIN_SC_CHOOSE_SERVER {
        ip?: (string|null);
        port?: (number|null);
        version?: (string|null);
        resVersion?: (string|null);
        resServerAddress?: (string|null);
        uid?: (number|null);
        token?: (string|null);
        serverId?: (number|null);
        serverName?: (string|null);
    }
// ...
    interface ILOGIN_SC_GET_SERVER_LIST {
        servers?: (S2C.LOGIN_SC_GET_SERVER_LIST.IServer[]|null);
    }
// ...
    namespace LOGIN_SC_GET_SERVER_LIST {

        interface IServer {
            serverId?: (number|null);
            name?: (string|null);
            status?: (number|null);
            level?: (number|null);
            lastLoginTime?: (number|null);
        }
// ...
    }

    interface ILOGIN_SC_GET_INFO {
        notice?: (string|null);
        version?: (string|null);
        updateAddress?: (string|null);
    }
// ...
    interface ILOGIN_SC_QUERY {
        ip?: (string|null);
        port?: (number|null);
        state?: (number|null);
        success?: (boolean|null);
    }
// ...
}